feat(admin blogs): add update blog process
+ add blog update route
+ add  blog update controller

* refactor multer and add conditions for skipping upload process if file was not uploaded

+++ app/utils/functions.js
@@ -154,7 +154,8 @@ function removeFile(fileAddress) {
     const fileAbsolutePath = path.join(__dirname, "..", "..", "public", fileAddress);
 
     /** remove file */
-    fs.unlinkSync(fileAbsolutePath);
+    if (fs.existsSync(fileAbsolutePath))
+        fs.unlinkSync(fileAbsolutePath);
 }
 
 module.exports = {
